feat(staff): recurring blocks, the schedule's grid views and the visit sheet's pages; the DS grows to match
Everything in progress on the branch since the bill ladder (2026-09-11),
and the block sheet's recurrence, rebuilt calendar-grade (2026-09-24).

Recurring blocks — "like Apple Calendar, MS Calendar, Google Calendar":
- domain/scheduling: `RecurrenceRule` and `RecurrenceSeries`, the RFC 5545
  subset the three calendars share — daily, weekly on a set of weekdays,
  monthly on a date or on the nth / last weekday, yearly; every N periods;
  until a day or N times; capped at 366 occurrences with typed refusals.
  Civil-date arithmetic on epoch days; missing dates skipped, never moved;
  the start bounds the series and occurs only if the rule selects it.
  No «Никога»: a block series is written day by day, so it must end.
- The block sheet: «Повтаряне» is a value row that pushes a page — the
  shortcuts (each saying what it means from this start), «Персонализирано»
  (frequency, a typed interval, the weekday toggles, the month's reading),
  how it ends, and the series read back; presets are derived from the rule,
  never stored. The commit carries the series' days.
- The write: a series is one batched pass per chair (`putSeries`: parallel
  reads, 500-write batches), not two round trips a day.
- DS: `ui-weekday-picker`; `ui-month-picker` marks days and takes a
  minimum; the end's calendar dots the series in secondary ink.
- The month picker, Apple's inline picker: a pale tint disc under a bold
  accent number at 20px for the chosen day (large text, so the contrast
  holds), the month as a capitalised title, tinted steps, an uppercase
  week key. Selected segments round to the pressed shape, wrapped or not.

The rest of the branch's work in progress:
- The staff schedule's grid views and their page stylesheet, the week and
  period labels in the day pill, view transitions, the shared event head.
- The visit sheet's pages: the add-client omnibox with local-first search
  and the tenant's dial code, the shop row, the money rows' pushed pages,
  photos with the camera, and one sheet-headline grammar for every page.
- DS: alert, avatar group, amount / code / search fields, choice tiles,
  segmented control, slider, scroll column, camera capture, code scanner,
  coupon card, sheet headline, and their showcase pages.
- Booking: the shop's own edits and placements are not bound by the
  catalogue's service conflicts; a client's coupons are readable to staff.
- Visit photos as staff-only sibling documents, with their rules.
- A session signs out only when the server says it is gone, and an
  emulator session survives a dev-stack restart.

+++ libs/features/shared/shell/src/lib/menu/site-menu.component.ts
@@ -158,7 +158,14 @@ export class SiteMenuComponent {
   );
 
   /**
-   * "сряда, 6 август" — the day the schedule opens on, as its own subtitle.
+   * "ср, 6 август" — the day the schedule opens on, as its own subtitle.
+   *
+   * The weekday SHORT (owner, 2026-09-17): «четвъртък, 17 септември» — the
+   * longest weekday and month Bulgarian has — outgrew the room the row
+   * leaves beside its mark on a 330px phone and wrapped to two lines,
+   * dropping the chevron under the date and growing the row. The short
+   * weekday reads as a date at a glance, can never wrap at any phone
+   * width, and is the weekday the staff day pill already uses.
    *
    * Deliberately a CLOCK READ, not a query. The obvious richer subtitle is
    * a live count ("4 visits today"), but the shop's day is N per-barber
@@ -173,7 +180,7 @@ export class SiteMenuComponent {
    */
   protected readonly todayLabel = computed(() =>
     new Intl.DateTimeFormat(this.lang() === 'en' ? 'en-GB' : 'bg-BG', {
-      weekday: 'long',
+      weekday: 'short',
       day: 'numeric',
       month: 'long',
       timeZone: 'Europe/Sofia',
